Refactor: delete Records class
