+++ src/caret_analyze/record/record_factory.py
@@ -20,7 +20,7 @@ from collections.abc import Sequence
 from multimethod import multimethod as singledispatchmethod
 
 from .column import ColumnValue
-from .record import Record, RecordInterface, Records, RecordsInterface
+from .record import Record, RecordInterface, RecordsInterface
 
 try:
     import caret_analyze.record.record_cpp_impl as cpp_impl
@@ -29,7 +29,7 @@ try:
     print('Succeed to find record_cpp_impl. the C++ version will be used.')
 except ModuleNotFoundError:
     use_cpp_impl = False
-    print('Failed to find record_cpp_impl. the Python version will be used.')
+    raise('Failed to find record_cpp_impl.')
 
 
 class RecordFactory:
